Added usage of BlockManager
The world is now loaded using the GetBlock method from the BlockManager instead of creating a block object on the heap for each block in the world.
Added BlocksByType to the BlockManager, which is a copy of the BlocksByIdentifier but mapped to their BlockType instead of their identifier
Added drop variable to the Block struct (and their json counterparts) which stores the block that should be dropped when the block has been broken.
Changed hardness to breaktime in the Block struct. The breaktime is now calculated when the blocks are read form the json using the hardness of each block.